backend/app/schemas/note_schema.py
from pydantic import BaseModel, Field
from typing import Optional
from bson import ObjectId

class PyObjectId(ObjectId):
    @classmethod
    def __get_validators__(cls):
        yield cls.validate

    @classmethod
    def validate(cls, v):
        if not ObjectId.is_valid(v):
            raise ValueError("Invalid ObjectId")
        return ObjectId(v)

class NoteCreate(BaseModel):
    content: str

class NoteUpdate(BaseModel):
    content: str

class NoteOut(BaseModel):
    id: str = Field(alias="_id")
    content: str
    owner_id: str

    class Config:
        allow_population_by_field_name = True
        json_encoders = {ObjectId: str}  