backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
motor
passlib[bcrypt]
pydantic[email] 
pyjwt

pytest
pytest-asyncio
httpx